Refactored Firebase imports
