+++ src/components/common/ProfilePanel.js
@@ -1,7 +1,8 @@
 import React from "react";
 import { Link } from "react-router-dom";
 import { Button } from "reactstrap";
-import firebase from "firebase";
+import * as firebase from "firebase/app";
+import "firebase/auth";
 import {
   MINIMUM_DISPLAY_NAME_LENGTH,
   MAXIMUM_DISPLAY_NAME_LENGTH,
